refactor(topBar): Split up topBar in container, bar and burgerMenu
